successfully refactored code into separate files
