+++ webpack.config.js
@@ -1,6 +1,7 @@
 const path = require('path');
 
 module.exports = {
+  // mode: 'development',
   entry: './src/index.js',
   output: {
     filename: 'bundle.js',
